+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- added retries to UCSC and NCBI
+- added retries to Travis tests
+
 ## [0.9.0] - 2020-09-01
 
 ### Added